requirements.txt
langchain
langchain-openai
faiss-cpu
python-dotenv
tiktoken
pypdf
streamlit